Introduce the basic scaffolding for a testing-only entry point in the Firebase Functions SDK.
The testing functions themselves live with the rest of the SDK's code, in src/. The src/testing.ts file is however also the m"main" for the testing/package.json, making it an access point reachable through the import "firebase-functions/testing".

The advantages of this approach are that generated code is still produced in outdir lib/, while the import path for testing-only methods remains appealing.

We stop packing the test/ directory (our internal unit tests) for distribution, to avoid user confusion between "test/" and "testing/".