Redesign SDK to get rid of "exports =" and export more types. (#82)
Redesign SDK to get rid of "exports =" and export more types.

Using "exports =" means that clients are supposed to need import =
require()[1]. This is a horrible anti-pattern that leaks to our
customers so we should avoid it.

This also tries to address outstanding concerns about how the SDK can be
made extensible for other providers and how we will make sure all types
customers need for production are exported in index.ts. To do this I
tried to follow a few principles:

1. A provider's export is a merged namespace + function. This styles the Functions SDK after the core JS SDK.
2. All provider namespaces should have a type called <provider>.FunctionBuilder. This improves predictability
3. Since providers' types will be referenced from their namespace, they should avoid a redundant prefix. E.g.
   database.FunctionBuilder not database.DatabaseFunctionBuilder.
4. All services offer a singleton at service() implementing service.Service. The singleton must be initialized
   in index.ts with service.init())
5. index.ts's purpose is to initialize services, rexport provider
   namespaces, and rexport loose/common types.

I admit the result is more like weak binding than true dependency injection because
the provider constructors depend on service singletons. The classes
should all be DI friendly still for better testability.

I'm increasingly uncomfortable with functions.env. Because env is a
property bag we can't actually export env.Env. There's really no
reasonable reason we would assume anyone can inject their own env type
for testing. I think we may have to make this functions.env() which
returns an env.Env (currently env.Data) or throws. I think env could be
its own conversation and I'm OK if we have to wait until the new year
for that.

In the meantime, the only functional change is that the failure case of
accessing functions.env too early changes from a helpful error message
to a null pointer exception.

[1]: https://www.typescriptlang.org/docs/handbook/modules.html