Introduce Firebase Analytics as an event provider (#121)
Introduces Firebase Analytics as an event provider in the Functions SDK.

The public API is:
> exports.myAnalyticsFunction = functions.analytics.event('analytics_event_type').onLog((event) => {
> ...
> });
This naming reflects the naming in the Analytics SDK.

Under the hood, quite a bit of code is spent on re-working the wire-format payload (a JSON-encoded version of the AnalyticsEvent proto) into a TS/JS-native object.

The code for this is placed in the constructors for the relevant types, which after much experimenting turned out to be the least invasive way of doing these transformations. We also use the nifty new "keyof" feature to prevent typos in field names.