Add a temporary shim to ignore spurious Storage events on function deployment. (#128)
Currently, running 'firebase deploy' will result in any Storage events triggering with a spurious (empty) event. This PR introduces a temporary shim in the SDK that will filter out such events, eliding the execution of the user's code.

This shim is sub-optimal, as any filtered events still count as an execution for purposes of UX, but it'll avoid the drawbacks of executing user code with invalid input. This shim will be removed when the backend has stopped sending these spurious events.