Set Analytics event.timestamp from event.data.logTime if timestamp is invalid (#129)
Currently, the Analytics backend is still sending an invalid (1969 or 1970) timestamp in 'event.timestamp'. This PR adds a shim to detect that, and pull a better timestamp from 'event.data.logTime', if available. This is temporary; we'll remove the shim when we roll a fix for the backend out to production.