Auth: reformat URL-formatted UIDs in the event payload. (#141)
Our backend is delivering URL-formatted UIDs (e.g. http://github.com/12345) for sign-ups from external event providers. This is out of spec, and out of sync with the admin SDK. We fix this here in the SDK so that customers will get the format fix when they next deploy. We'll fix the format sent by the backend further into the future; this SDK fix is forwards-compatible for that.