Add an extremely basic integration test to the SDK (#150)
Add an extremely basic integration test to the SDK, that SDK developers can run against their own Firebase projects to check that all moving parts still work in practice.

Previously, many developers had something like this stashed somewhere on their machine. The hope is that by adding this to the Git repo we'll maintain the integration test in a single place as we add new SDK features.