Moves most of https pacakge into /common/providers/https. (#915)
Moves most of https pacakge into /common/providers/https.

The actual encoding/decoding/error logic etc. of callable functions
will stay the same between v1 and v2, so they should be in a /common
file.

The rest of /v1 is about trigger options, cors, etc.

onCall's CORS options are part of the v1 namespace because v2 adds
a "cors" option to HttpsOptions. In the v2 API, we will default
to `"cors": true`, but a customer could say `"cors": "my.firebase.app"`
and we will respsect it.

I wish there was a way to make this diff not so noisy. Most of this
is cut & paste.