Use npm ci on publish scripts. (#940)
I kept seeing error on the npm version command claiming that the directory was dirty. I believe this is because version of the npm used in our builder image is >7 which uses lockfileversion 2 instead of 1 checked into the repo. Using npm ci instead, but we should also think about updating the package-lock.json to lockfileversion 2.