Container contract support for Firebase Functions SDK (#1014)
* Annotate function triggers with __endpoint property (#999)

In addition to annotating function triggers with `__trigger` property, we add `__endpoint` annotation. This property will be used by the to-be-developed functions runtime to generate/declare deployment manifest that the CLI will use to deploy the function.

There are lots of code duplication between the utility functions for annotating the `__trigger` and `__endpoint` properties. I didn't try to refactor the common code since I expect that we will favor `__endpoint` property in the future.

* Add missing import.

* More of annotate endpoint property for functions (#1009)

Follows up https://github.com/firebase/firebase-functions/pull/999 to annotate each funuctions with `__endpoint` property.

Highlight of changes:

* Extend unit test coverage for all v1 providers
* Add `__endpoint` annotation to v1 task queue functions
* Add `__requiredAPIs` annotation to task queue and scheduler functions
* Explicitly set `__endpoint` to undefined in the handler namespace
* No SDK-level label setting in the __endpoint annotation.