Make event params strongly typed (#1155)
Implements go/cf3-typed-params.

WIP: currently only common and v2 are implemented.

Had to update `prettier` to the next major version so it could read TS 4.1 features without erroring out. I created  #1160 as a base for this change to make it easier to separate mechanical from hand-crafted changes. If this is approved, however, I'll commit both changes to `launch.next`