Adds 'eventarcpublishing' as required API to custom event function (#1300)
Adds `eventarcpublishing.googleapis.com` as a required API for the `onCustomEventPublished` trigger so that the CLI enables it on deployment. Attempting to deploy a cloud function to GCF that listens on a non-existent channel will throw an error; `eventarcpublishing` API is necessary to allow the CLI to create the channel before the function is deployed.